showTeamModal shows the team members as a table correctly
