+++ src/assignments/components/ShowTeamModal.tsx
@@ -1,7 +1,9 @@
 import { useState } from "react"
 import Modal from "src/core/components/Modal"
 import { getContributorName, getTeamName } from "src/services/getName"
-import AssignTeamMembers, { TeamOption } from "src/teams/components/TeamMembersTable"
+import { Tooltip } from "react-tooltip"
+import { TeamMembers, teamMembersTableColumns } from "src/teams/components/TeamMembersTable"
+import Table from "src/core/components/Table"
 
 export const ShowTeamModal = ({ team }) => {
   const [openModal, setOpenModal] = useState(false)
@@ -10,11 +12,8 @@ export const ShowTeamModal = ({ team }) => {
 
   const teamMembers = contributors.map((contributor) => {
     return {
-      userName: getContributorName(contributor),
-      id: contributor.id,
-      checked: false,
-      teamId: team.id,
-    } as TeamOption
+      username: getContributorName(contributor),
+    } as TeamMembers
   })
 
   // Handle events
@@ -23,24 +22,28 @@ export const ShowTeamModal = ({ team }) => {
   }
 
   return (
-    <>
-      <div className="mt-4">
-        <button onClick={() => handleToggle()}>
-          <span>{`${getTeamName(team)}`}</span>
-        </button>
+    <div className="mt-4">
+      <button
+        type="button"
+        className="btn btn-primary"
+        data-tooltip-id="showTeamModalTooltip"
+        onClick={() => handleToggle()}
+      >
+        <span>{`${getTeamName(team)}`}</span>
+      </button>
+      <Tooltip id="showTeamModalTooltip" content="Show team members" className="z-[1099]" />
 
-        <Modal open={openModal} size="w-7/6 max-w-1xl">
-          <div className="flex justify-start mt-4">
-            <AssignTeamMembers showCheckbox={false} teamOptions={teamMembers}></AssignTeamMembers>
-          </div>
-          <div className="modal-action flex justify-end mt-4">
-            {/* TODO: Add tooltip that say: click to show team members */}
-            <button type="button" className="btn btn-primary" onClick={handleToggle}>
-              Close
-            </button>
-          </div>
-        </Modal>
-      </div>
-    </>
+      <Modal open={openModal} size="w-7/6 max-w-1xl">
+        <div className="flex flex-col justify-start mt-4">
+          <h3>Team members</h3>
+          <Table columns={teamMembersTableColumns} data={teamMembers} />
+        </div>
+        <div className="modal-action flex justify-end mt-4">
+          <button type="button" className="btn btn-primary" onClick={handleToggle}>
+            Close
+          </button>
+        </div>
+      </Modal>
+    </div>
   )
 }
